fix: Remove commit hash from package versions in generated definition projects
Commit part is only included if Generation project is built with .NET SDK 8.0 or higher.

This is due to the breaking change of .NET 8 SDK:

```
The Source Link build tooling is now included in the .NET SDK.
Source Link enables packages and applications to embed information about the source control information of the built artifacts.
As a side effect, commit information is included in the InformationalVersion value of built libraries and applications.

This change affects any project that's built with the .NET 8 SDK, even those that target .NET 7 or an earlier version.

Prior to this change, the default InformationalVersion of a library or application was the Version property.
Starting in .NET 8, the default InformationalVersion of a library or application is the Version property and the SourceRevisionId property.
```